Rename schedulers module to coordinators
Move schedulers.py to coordinators.py and update all imports.

Keep coordinator behavior unchanged while aligning module naming with class names.
