+++ tests/unit/test_coordinator.py
@@ -10,12 +10,12 @@ from homeassistant.helpers import frame
 from homeassistant.helpers.update_coordinator import UpdateFailed
 
 from custom_components.mittfortum.api.client import FortumAPIClient
-from custom_components.mittfortum.exceptions import APIError
-from custom_components.mittfortum.models import ConsumptionData
-from custom_components.mittfortum.schedulers import (
+from custom_components.mittfortum.coordinators import (
     HourlyConsumptionSyncCoordinator,
     SpotPriceSyncCoordinator,
 )
+from custom_components.mittfortum.exceptions import APIError
+from custom_components.mittfortum.models import ConsumptionData
 
 
 @pytest.fixture
