+++ tests/unit/test_entity.py
@@ -4,9 +4,9 @@ from unittest.mock import Mock
 
 import pytest
 
+from custom_components.mittfortum.coordinators import HourlyConsumptionSyncCoordinator
 from custom_components.mittfortum.device import MittFortumDevice
 from custom_components.mittfortum.entity import MittFortumEntity
-from custom_components.mittfortum.schedulers import HourlyConsumptionSyncCoordinator
 
 
 @pytest.fixture
